separate and extract pumping widgetbook use cases and mocking its context
