+++ packages/widgetbook_golden_test/pubspec.yaml
@@ -22,7 +22,6 @@ dependencies:
 
 dev_dependencies:
   flutter_lints: ^5.0.0
-  mocktail_image_network: ^1.2.0
   example:
     path: example
   provider: ^6.1.5+1
